Internal: fixes to grid-size-field [ED-23845]
